+++ lib/src/widgets/inputs/mp_azimuth_picker_widget.dart
@@ -70,7 +70,7 @@ class _MPAzimuthPickerWidgetState extends State<MPAzimuthPickerWidget> {
   double _calculateAngle({required Offset center, required Offset position}) {
     final Offset delta = position - center;
     // Convert cartesian to polar coordinates (with y inverted)
-    final double angle = math.atan2(delta.dx, -delta.dy) * mp1Radian;
+    final double angle = math.atan2(delta.dx, -delta.dy) * mp1RadInDegree;
 
     return angle;
   }